pyproject.toml
[build-system]
requires = ["setuptools>=66"]
build-backend = "setuptools.build_meta"

[project]
name = "agsekit"
version = "1.1.2"
description = "Agent Safety Kit command-line utilities"
authors = [
  { name = "Mihanentalpo", email = "mihanentalpo@yandex.ru" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
  "ansible-core>=2.16,<2.19; python_version >= '3.10'",
  "ansible-core>=2.15,<2.16; python_version < '3.10'",
  "click>=8.1,<9",
  "PyYAML>=6.0,<7",
  "questionary>=2.0,<3",
  "ruamel.yaml>=0.18,<0.19",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.scripts]
agsekit = "agsekit_cli.cli:main"

[tool.setuptools.packages.find]
include = ["agsekit_cli", "agsekit_cli.*"]

[tool.setuptools.package-data]
agsekit_cli = [
  "agent_scripts/*.sh",
  "agent_scripts/proxychains_common.sh",
  "ansible/*.yml",
  "ansible/agents/*.yml",
  "ansible/bundles/*.yml",
  "vm_installers/*.sh",
  "run_with_proxychains.sh",
  "locales/*.json",
]

[tool.setuptools.data-files]
"share/agsekit" = ["config-example.yaml"]
